+++ packages/api/src/hooks/useSuggestedActions.ts
@@ -1,5 +1,10 @@
 import { useSuggestedActionsHooks } from 'botframework-webchat-redux-store';
 
+/**
+ * @deprecated Use `useSuggestedActionsHooks().useSuggestedActions` instead. This hook will be removed on or after 2027-05-30.
+ */
 export default function useSuggestedActions() {
+  // Provides a path for backward compatibility during deprecation.
+  // eslint-disable-next-line local-rules/forbid-use-hook-producer
   return useSuggestedActionsHooks().useSuggestedActions();
 }